backend/requirements.txt
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
Django==5.2.12
djangorestframework==3.17.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.5
typing_extensions==4.15.0
gunicorn
